refactor(config): split AppConfig by domain (Phase 4)
Implements Phase 4 of the Rust structure refactor roadmap. Splits the
1798-LOC `src/config.rs` into five domain-focused files with
`AppConfig` staying as the persistence boundary and its `impl` block
distributed across submodules by concern.

Split:

- `persist.rs` (78 LOC) — `load_or_init` + `save`. File I/O and
  bootstrap plumbing.
- `agents.rs` (446 LOC) — `resolve_agent_source`,
  `resolve_auth_forward_mode`, `set_agent_auth_forward`,
  `trust_agent`, `untrust_agent`, `is_builtin_agent`,
  `sync_builtin_agents`, and the `BUILTIN_AGENTS` registry.
- `mounts.rs` (372 LOC) — `resolve_mounts`,
  `expand_and_validate_named_mounts`, `add_mount`, `remove_mount`,
  `list_mounts`, `global_mounts`, plus the `MountEntry` enum and
  `DockerMounts` wrapper + its methods.
- `workspaces.rs` (345 LOC) — `create_workspace`, `edit_workspace`,
  `remove_workspace`, `list_workspaces`, `insert_workspace_raw`,
  `validate_workspaces`.
- `mod.rs` (601 LOC, down from 1798) — the data types (`AppConfig`,
  `AuthForwardMode`, `ClaudeConfig`, `ClaudeAgentConfig`,
  `AgentSource`, `DockerConfig`), module declarations, and re-exports
  of `DockerMounts` / `MountEntry`. Plus 15 cross-cutting tests that
  span more than one domain.

Verbatim moves. No function body edits. No logic changes.

Test Preservation Policy honored: 418 tests preserved, co-located:

- `persist.rs` — 1 test (bootstrap/idempotency round-trip).
- `agents.rs` — 16 tests (builtin sync, trust, auth-forward).
- `mounts.rs` — 10 tests (docker mount deserialization,
  `resolve_mounts`, `validate_mounts_*`).
- `workspaces.rs` — 5 tests (edit/create/remove validation).
- `mod.rs` — 15 tests that span domains (held in `mod.rs` rather
  than duplicating setup across submodules).

Visibility changes, all bounded to `pub(super)` (inside `config`):
- `BUILTIN_AGENTS` constant (was private).
- `sync_builtin_agents` (was private) — needed by `persist::load_or_init`.
- `validate_workspaces` (was private) — needed by `persist::load_or_init`.

No public API widened. All external users
(`src/app/mod.rs`, `src/app/context.rs`, `src/runtime.rs`,
`src/launch.rs`, `src/instance.rs`, `src/workspace/resolve.rs`,
`tests/workspace_config_crud.rs`) continue to import
`crate::config::{AppConfig, AgentSource, MountConfig, MountEntry,
AuthForwardMode, DockerMounts}` unchanged.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
