+++ src/config/persist.rs
@@ -1,0 +1,78 @@
+use super::AppConfig;
+use crate::paths::JackinPaths;
+
+impl AppConfig {
+    pub fn load_or_init(paths: &JackinPaths) -> anyhow::Result<Self> {
+        paths.ensure_base_dirs()?;
+
+        let mut config = match std::fs::read_to_string(&paths.config_file) {
+            Ok(contents) => toml::from_str(&contents)?,
+            Err(e) if e.kind() == std::io::ErrorKind::NotFound => Self::default(),
+            Err(e) => return Err(e.into()),
+        };
+
+        if config.sync_builtin_agents() {
+            config.save(paths)?;
+        }
+
+        config.validate_workspaces()?;
+        Ok(config)
+    }
+
+    pub fn save(&self, paths: &JackinPaths) -> anyhow::Result<()> {
+        let contents = toml::to_string_pretty(self)?;
+        let tmp = paths.config_file.with_extension("tmp");
+
+        #[cfg(unix)]
+        {
+            use std::io::Write;
+            use std::os::unix::fs::OpenOptionsExt;
+            let mut file = std::fs::OpenOptions::new()
+                .write(true)
+                .create(true)
+                .truncate(true)
+                .mode(0o600)
+                .open(&tmp)?;
+            file.write_all(contents.as_bytes())?;
+            file.sync_all()?;
+        }
+
+        #[cfg(not(unix))]
+        std::fs::write(&tmp, &contents)?;
+
+        std::fs::rename(&tmp, &paths.config_file)?;
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::paths::JackinPaths;
+    use tempfile::tempdir;
+
+    #[test]
+    fn sync_does_not_rewrite_config_when_already_current() {
+        let temp = tempdir().unwrap();
+        let paths = JackinPaths::for_tests(temp.path());
+
+        // First load creates the file
+        AppConfig::load_or_init(&paths).unwrap();
+        let mtime_before = std::fs::metadata(&paths.config_file)
+            .unwrap()
+            .modified()
+            .unwrap();
+
+        // Small delay so mtime would differ if rewritten
+        std::thread::sleep(std::time::Duration::from_millis(50));
+
+        // Second load should not rewrite
+        AppConfig::load_or_init(&paths).unwrap();
+        let mtime_after = std::fs::metadata(&paths.config_file)
+            .unwrap()
+            .modified()
+            .unwrap();
+
+        assert_eq!(mtime_before, mtime_after);
+    }
+}
